@@ -1,49 +1,116 @@
 {
-	"name": "3brs/sylius-payment-fee-plugin",
-	"type": "sylius-plugin",
-	"description": "Payment fee plugin for Sylius.",
-	"license": "MIT",
-	"require": {
-		"php": "^7.3|^8.0",
-		"sylius/sylius": "1.7.*|1.8.*|1.9.*|1.10.*"
-	},
-	"require-dev": {
-		"behat/behat": "^3.6.1",
-		"behat/mink-selenium2-driver": "^1.4",
-		"dmore/behat-chrome-extension": "^1.3",
-		"dmore/chrome-mink-driver": "^2.7",
-		"friends-of-behat/mink": "^1.8",
-		"friends-of-behat/mink-browserkit-driver": "^1.4",
-		"friends-of-behat/mink-extension": "^2.4",
-		"friends-of-behat/page-object-extension": "^0.3",
-		"friends-of-behat/suite-settings-extension": "^1.0",
-		"friends-of-behat/symfony-extension": "^2.1",
-		"friends-of-behat/variadic-extension": "^1.3",
-		"friendsofsymfony/oauth-server-bundle": "^1.6 || >2.0.0-alpha.0 ^2.0@dev",
-		"lakion/mink-debug-extension": "^2.0.0",
-		"phpspec/phpspec": "^7.0",
-		"phpstan/phpstan": "0.12.88",
-		"phpstan/phpstan-symfony": "^0.12.7",
-		"phpstan/phpstan-doctrine": "0.12.37",
-		"phpunit/phpunit": "^8.5",
-		"sensiolabs/security-checker": "^6.0",
-		"sylius-labs/coding-standard": "^4.0.0",
-		"symfony/browser-kit": "^4.4|^5.2",
-		"symfony/debug-bundle": "^4.4|^5.2",
-		"symfony/dotenv": "^4.4|^5.2",
-		"symfony/intl": "^4.4|^5.2",
-		"symfony/web-profiler-bundle": "^4.4|^5.2"
-	},
-	"prefer-stable": true,
-	"autoload": {
-		"psr-4": {
-			"ThreeBRS\\SyliusPaymentFeePlugin\\": "src/",
-			"Tests\\ThreeBRS\\SyliusPaymentFeePlugin\\": "tests/"
-		}
-	},
-	"autoload-dev": {
-		"psr-4": {
-			"App\\": "tests/Application/src/"
-		}
-	}
+  "name": "3brs/sylius-payment-fee-plugin",
+  "type": "sylius-plugin",
+  "description": "Payment extra fee plugin for Sylius",
+  "keywords": [
+    "sylius",
+    "sylius-plugin",
+    "plugin",
+    "payment",
+    "extra",
+    "fee"
+  ],
+  "license": "MIT",
+  "require": {
+    "php": "^8.0",
+    "sylius/sylius": "^1.11|^1.12"
+  },
+  "require-dev": {
+    "behat/behat": "^3.9.0",
+    "behat/mink-selenium2-driver": "^1.4",
+    "dmore/behat-chrome-extension": "^1.3",
+    "dmore/chrome-mink-driver": "^2.7",
+    "friends-of-behat/mink": "^1.10",
+    "friends-of-behat/mink-browserkit-driver": "^1.4",
+    "friends-of-behat/mink-debug-extension": "^2.1",
+    "friends-of-behat/mink-extension": "^2.4",
+    "friends-of-behat/page-object-extension": "^0.3",
+    "friends-of-behat/suite-settings-extension": "^1.0",
+    "friends-of-behat/symfony-extension": "^2.1",
+    "friends-of-behat/variadic-extension": "^1.3",
+    "php-http/message-factory": "^1.0.0",
+    "phpstan/phpstan": "^1.8.1",
+    "phpstan/phpstan-doctrine": "1.3.37",
+    "phpstan/phpstan-strict-rules": "^1.3.0",
+    "phpstan/phpstan-symfony": "^1.3",
+    "phpstan/phpstan-webmozart-assert": "^1.2.0",
+    "phpunit/phpunit": "^9.5",
+    "polishsymfonycommunity/symfony-mocker-container": "^1.0",
+    "rector/rector": "^0.18.12",
+    "sylius-labs/coding-standard": "^4.2",
+    "symfony/browser-kit": "^5.4 || ^6.0",
+    "symfony/debug-bundle": "^5.4 || ^6.0",
+    "symfony/dotenv": "^5.4 || ^6.0",
+    "symfony/flex": "^2.2.2",
+    "symfony/framework-bundle": "^5.4 || ^6.0",
+    "symfony/http-foundation": "^5.4 || ^6.0",
+    "symfony/http-kernel": "^5.4 || ^6.0",
+    "symfony/intl": "^5.4 || ^6.0",
+    "symfony/web-profiler-bundle": "^5.4 || ^6.0",
+    "vimeo/psalm": "5.9.0"
+  },
+  "conflict": {
+    "babdev/pagerfanta-bundle": "<3.6.1",
+    "doctrine/collections": "<1.7.0",
+    "doctrine/dbal": "<2.13.3",
+    "doctrine/doctrine-bundle": "<2.4.2",
+    "friendsofsymfony/rest-bundle": "<3.1.0",
+    "jms/serializer-bundle": "<4.2.0",
+    "masterminds/html5": "<2.7.5",
+    "knplabs/knp-menu": "<3.3.0",
+    "lexik/jwt-authentication-bundle": "<2.12",
+    "polishsymfonycommunity/symfony-mocker-container": "<1.0.6",
+    "sylius/resource-bundle": "<1.10.0",
+    "symfony/css-selector": "<4.4.24",
+    "symfony/framework-bundle": ">=6.1.0 <6.1.4|6.2.8",
+    "symfony/mime": "<5.4.0",
+    "symfony/web-link": "<5.3.0",
+    "symplify/easy-coding-standard": "<10.2.0",
+    "twig/twig": "<2.14.7",
+    "webmozart/assert": "<1.11.0",
+    "willdurand/negotiation": "<3.0"
+  },
+  "prefer-stable": true,
+  "config": {
+    "sort-packages": true,
+    "allow-plugins": {
+      "symfony/flex": true,
+      "dealerdirect/phpcodesniffer-composer-installer": true
+    }
+  },
+  "extra": {
+    "symfony": {
+      "require": "^5.4 || ^6.0"
+    }
+  },
+  "autoload": {
+    "psr-4": {
+      "ThreeBRS\\SyliusPaymentFeePlugin\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Tests\\ThreeBRS\\SyliusPaymentFeePlugin\\": "tests/Application/src/",
+      "Tests\\ThreeBRS\\SyliusPaymentFeePlugin\\Behat\\": "tests/Behat/"
+    }
+  },
+  "exclude-from-classmap": [
+    "/vendor/sylius/grid-bundle/src/Bundle/test/"
+  ],
+  "scripts": {
+    "post-install-cmd": [
+      "php bin/create_node_symlink.php"
+    ],
+    "post-update-cmd": [
+      "php bin/create_node_symlink.php"
+    ],
+    "post-create-project-cmd": [
+      "php bin/create_node_symlink.php"
+    ],
+    "auto-scripts": {
+      "cache:clear": "symfony-cmd",
+      "assets:install %PUBLIC_DIR%": "symfony-cmd",
+      "security-checker security:check": "script"
+    }
+  }
 }
