@@ -1,20 +1,20 @@
 module.exports = {
-    extends: 'airbnb-base',
-    env: {
-        node: true,
-    },
-    rules: {
-        'object-shorthand': ['error', 'always', {
-            avoidQuotes: true,
-            avoidExplicitReturnArrows: true,
-        }],
-        'function-paren-newline': ['error', 'consistent'],
-        'max-len': ['warn', 120, 2, {
-            ignoreUrls: true,
-            ignoreComments: false,
-            ignoreRegExpLiterals: true,
-            ignoreStrings: true,
-            ignoreTemplateLiterals: true,
-        }],
-    },
+  extends: 'airbnb-base',
+  env: {
+    node: true,
+  },
+  rules: {
+    'object-shorthand': ['error', 'always', {
+      avoidQuotes: true,
+      avoidExplicitReturnArrows: true,
+    }],
+    'function-paren-newline': ['error', 'consistent'],
+    'max-len': ['warn', 120, 2, {
+      ignoreUrls: true,
+      ignoreComments: false,
+      ignoreRegExpLiterals: true,
+      ignoreStrings: true,
+      ignoreTemplateLiterals: true,
+    }],
+  },
 };
